+++ job_runner/CHANGELOG.md
@@ -14,6 +14,8 @@ https://github.com/ai2cm/full-model/compare/job_runner
   priority balancer ([#1419](https://github.com/ai2cm/ace/pull/1419)). It
   defaults to the `priority` field from the input file and can be overridden for
   a whole invocation with the `CM_PRIORITY` environment variable.
+- [FEAT] Add `--config-dir` arg to `{evaluate,inference}.sh` for shared eval
+  configs
 
 ### Changed
 
@@ -77,5 +79,3 @@ https://github.com/ai2cm/full-model/compare/job_runner
 - [CHORE] Automate porting of experiment-branch-to-job_runner-branch updates
 - [ISSUE] `make jr_changelog` doesn't properly handle insertion after multi-line
   items
-- [FEAT] Add `--config-dir` arg to `{evaluate,inference}.sh` for shared eval
-  configs